move axios to components/ and create some files because of auth token
