+++ app/frontend/components/axios/interceptors.jsx
@@ -1,0 +1,21 @@
+export const reqAuth = (instance) => {
+  instance.interceptors.request.use(
+    (config) => {
+      return config
+    },
+    (error) => {
+      return Promise.reject(error)
+    }
+  )
+}
+
+export const resAuth = (instance) => {
+  instance.interceptors.request.use(
+    (config) => {
+      return config
+    },
+    (error) => {
+      return Promise.reject(error)
+    }
+  )
+}
